requirements.txt
appnope==0.1.3
asttokens==2.4.1
certifi==2021.5.30
cffi==1.16.0
chardet==4.0.0
charset-normalizer==3.3.2
comm==0.2.0
cryptography==42.0.3
customtkinter==5.2.2
darkdetect==0.8.0
debugpy==1.8.0
decorator==5.1.1
distlib==0.3.8
executing==2.0.1
falcon==3.0.1
filelock==3.13.4
idna==2.10
ipykernel==6.27.1
ipython==8.18.1
jedi==0.19.1
jupyter_client==8.6.0
jupyter_core==5.5.0
matplotlib-inline==0.1.6
mysql-connector-python==8.2.0
mysql-connector-python-rf==2.2.2
nest-asyncio==1.5.8
packaging==23.2
parso==0.8.3
pdfminer.six==20231228
pexpect==4.9.0
pipenv==2023.12.1
platformdirs==4.1.0
prompt-toolkit==3.0.41
protobuf==4.21.12
psutil==5.9.6
ptyprocess==0.7.0
pure-eval==0.2.2
pycparser==2.21
Pygments==2.17.2
python-dateutil==2.8.2
pyzmq==25.1.2
requests==2.25.1
setuptools==69.5.1
six==1.16.0
somepackage==1.2.3
stack-data==0.6.3
tk==0.1.0
tornado==6.4
traitlets==5.14.0
urllib3==1.26.6
virtualenv==20.26.0
wcwidth==0.2.12